Title:fix,the TaskDialog not content when delete a lot of files.
Description:bug-26944,【5.1.1.39-1】删除大量文件夹的进度条展示不正确

RootCause:以前该功能未实现

Solution:实现该功能

+++ dialogs/dtaskdialog.cpp
@@ -653,7 +653,7 @@ void DTaskDialog::updateData(DFMTaskWidget *wid, const QMap<QString, QString> &d
     }
 
     if (!progress.isEmpty()) {
-        wid->onProgressChanged(progress.toInt(), 0);
+        wid->onProgressChanged(progress.toDouble(), 0);
     }
 }
